Remove overreaching clause.
We mustn’t uniquely vilify sexual content. That content which is harassing or
otherwise problematic to the project or its members is prohibited by the rest
of the CC. While sexual content probably has no place in a programming language,
proactively forbidding it smacks of a sort of misguided moralistic overreach.